refactor(reorder-list): renamed examples/showcase files and folders
