+++ src/app/examples/examples.routes.ts
@@ -47,7 +47,7 @@ import { ListColoredItemsExampleComponent } from './list/components/colored-item
 import { ListWithHeaderAndFooterExampleComponent } from './list/components/header-and-footer';
 import { ListWithSectionsExampleComponent } from './list/components/sections';
 import { ListItemsExampleComponent } from './list/components/items';
-import { ReorderListExampleComponent } from './reorder/reorder-list-example.component';
+import { ReorderListExampleComponent } from './reorder-list/reorder-list-example.component';
 import { DropdownExampleComponent } from '~/app/examples/dropdown-example/dropdown-example.component';
 
 export const routes: Routes = [
