+++ src/app/page/side-nav/side-nav.component.html
@@ -56,7 +56,7 @@
       <li><a routerLink="showcase/list-load-on-demand" [routerLinkActive]="'is-selected'" class="side-nav__item">List - load on demand</a></li>
       <li><a routerLink="showcase/list-swipe" [routerLinkActive]="'is-selected'" class="side-nav__item">List - swipe</a></li>
       <li><a routerLink="showcase/list-no-shape" [routerLinkActive]="'is-selected'" class="side-nav__item">List - No styling</a></li>
-      <li><a routerLink="showcase/reorder" [routerLinkActive]="'is-selected'" class="side-nav__item">List - reorder</a></li>
+      <li><a routerLink="showcase/reorder-list" [routerLinkActive]="'is-selected'" class="side-nav__item">List - reorder</a></li>
       <li><a routerLink="showcase/modal" [routerLinkActive]="'is-selected'" class="side-nav__item">Modal</a></li>
       <li><a routerLink="showcase/loading-overlay" [routerLinkActive]="'is-selected'" class="side-nav__item">Loading Overlay</a></li>
       <li><a routerLink="showcase/segmented-control" [routerLinkActive]="'is-selected'" class="side-nav__item">Segmented Control</a></li>
