+++ src/app/showcase/showcase.common.ts
@@ -1,4 +1,3 @@
-import { ReorderShowcaseComponent } from './reorder-showcase/reorder-showcase.component';
 import { FormFieldShowcaseComponent } from './form-field-showcase/form-field-showcase.component';
 import { ListNoShapeShowcaseComponent } from './list-no-shape-showcase/list-no-shape-showcase.component';
 import { ListSwipeShowcaseComponent } from './list-swipe-showcase/list-swipe-showcase.component';
@@ -36,6 +35,7 @@ import { TabsShowcaseComponent } from './tabs-showcase/tabs-showcase.component';
 import { ItemShowcaseComponent } from '../showcase/item-showcase/item-showcase.component';
 import { DividerShowcaseComponent } from '../showcase/divider-showcase/divider-showcase.component';
 import { DropdownShowcaseComponent } from '~/app/showcase/dropdown-showcase/dropdown-showcase.component';
+import { ReorderListShowcaseComponent } from '../showcase/reorder-list-showcase/reorder-list-showcase.component';
 
 export const COMPONENT_IMPORTS: any[] = [ExamplesModule, ShowcaseRoutingModule];
 
@@ -73,7 +73,7 @@ export const COMPONENT_EXPORTS: any[] = [
   TabsShowcaseComponent,
   ItemShowcaseComponent,
   DividerShowcaseComponent,
-  ReorderShowcaseComponent,
+  ReorderListShowcaseComponent,
   DropdownShowcaseComponent,
 ];
