+++ src/app/showcase/showcase.routes.ts
@@ -34,8 +34,8 @@ import { PageShowcaseComponent } from './page-showcase/page-showcase.component';
 import { TabsShowcaseComponent } from './tabs-showcase/tabs-showcase.component';
 import { ItemShowcaseComponent } from '../showcase/item-showcase/item-showcase.component';
 import { DividerShowcaseComponent } from './divider-showcase/divider-showcase.component';
-import { ReorderShowcaseComponent } from './reorder-showcase/reorder-showcase.component';
 import { DropdownShowcaseComponent } from '~/app/showcase/dropdown-showcase/dropdown-showcase.component';
+import { ReorderListShowcaseComponent } from './reorder-list-showcase/reorder-list-showcase.component';
 
 export const routes: Routes = [
   {
@@ -180,8 +180,8 @@ export const routes: Routes = [
         component: FormFieldShowcaseComponent,
       },
       {
-        path: 'reorder',
-        component: ReorderShowcaseComponent,
+        path: 'reorder-list',
+        component: ReorderListShowcaseComponent,
       },
       {
         path: 'dropdown',
